requirements.txt
Flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
sentence-transformers>=2.2.2
faiss-cpu>=1.13.0
numpy>=1.26.0

